app/requirements.txt
opencv-python>=4.10.0.8
opencv-contrib-python>=4.10.0.84
mss>=10.0.0
pillow>=11.1.0